# Faster implementation for constructing gradient of extract_image_patches