Fix #20146.

In the PR, we created indices matrix for input tensor, and reuse `extract_image_patches` to calculate `idx` (the mapping table for indices: input -> output). 

The list below is the performance comparison based on the example of @dbbert in #20146.

Before:
+ constructing time: ~138.4s
+ execute time: ~33.7s

After:
+ constructing time: ~3.6s
+ execute time: ~20.3s