@facaiy 2018-07-29
Test script:

```python
import tensorflow as tf
import time

tf.reset_default_graph()

batch_size = 4
height = width = 1024
ksize = 5

inputs = tf.get_variable('inputs', (batch_size, height, width, 1))
print('inputs: ', inputs)
patches = tf.extract_image_patches(inputs, 
                                   ksizes=[1,ksize,ksize,1], 
                                   strides=[1,1,1,1], 
                                   rates=[1,1,1,1], 
                                   padding='SAME')
print('patches: ', patches.shape)
s = time.clock()
gradients = tf.gradients(patches, inputs)
print('construct time: {}'.format(time.clock() - s))

with tf.Session() as sess:
    print('gradients: ', gradients)
    s = time.clock()
    res = sess.run(gradients)
    print('exec time: {}'.format(time.clock() - s))
    print('res shape: ', res[0].shape)
    print('res: ', res[0][0, 0:3, 0:3, 0])
```