@alextp 2018-08-13
Can you also add this benchmark to a test, so we can track future regressions?