@facaiy 2018-08-14
Sure, could you take a look again? Because execute time is too long (> 20s), we only add benchmark case for constructing gradient here.